pyproject.toml
[project]
name = "langgraph-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.0.4",
    "langchain-openai>=0.3.8",
    "langgraph-cli[inmem]>=0.1.77",
    "rich>=13.9.4",
    "trustcall>=0.0.38",
]
